@@ -1,18 +1,15 @@
 {
   "name": "webpack-remove-empty-scripts",
-  "version": "1.0.3",
-  "description": "Plugin for Webpack 5 to remove empty JavaScript files generated when using only styles in Webpack entry.",
+  "version": "1.0.4",
+  "description": "Webpack plugin removes empty JavaScript files generated when using styles.",
   "keywords": [
     "webpack",
+    "plugin",
     "remove",
     "empty",
     "script",
-    "mini",
-    "extract",
-    "plugin",
     "entry",
     "style",
-    "scss",
     "scss",
     "sass",
     "css",
@@ -43,6 +40,7 @@
   },
   "main": "src/index.js",
   "scripts": {
+    "install:yarn": "yarn install --ignore-engines",
     "test": "jest --detectOpenHandles --config ./test/jest.config.js",
     "test:coverage": "jest --detectOpenHandles --collectCoverage --config ./test/jest.config.js",
     "publish:public": "npm publish --access public",
@@ -64,20 +62,21 @@
     "ansis": "1.5.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.21.4",
-    "@babel/preset-env": "^7.21.4",
-    "@test-fixtures/lorem" : "0.0.2",
-    "@types/jest": "^28.1.8",
-    "@wordpress/scripts": "^23.7.2",
-    "@wordpress/dependency-extraction-webpack-plugin": "^4.13.0",
-    "css-loader": "^6.7.3",
+    "@babel/core": "^7.22.11",
+    "@babel/preset-env": "^7.22.10",
+    "@test-fixtures/lorem": "0.0.2",
+    "@types/jest": "^29.5.4",
+    "@wordpress/scripts": "^26.11.0",
+    "@wordpress/dependency-extraction-webpack-plugin": "^4.23.0",
+    "css-loader": "^6.8.1",
+    "html-bundler-webpack-plugin": "^2.9.0",
     "html-loader": "^4.2.0",
-    "jest": "^28.1.3",
-    "mini-css-extract-plugin": "^2.7.5",
-    "terser-webpack-plugin": "^5.3.7",
-    "webpack": "^5.78.0",
-    "webpack-hot-middleware": "^2.25.3",
+    "jest": "^29.6.4",
+    "mini-css-extract-plugin": "^2.7.6",
+    "terser-webpack-plugin": "^5.3.9",
+    "webpack": "^5.88.2",
+    "webpack-hot-middleware": "^2.25.4",
     "webpack-manifest-plugin": "^5.0.0",
-    "webpack-merge": "^5.8.0"
+    "webpack-merge": "^5.9.0"
   }
 }
