@@ -4,7 +4,7 @@ const RemoveEmptyScriptsPlugin = require('../../../src/index.js');
 module.exports = {
   mode: 'production',
   output: {
-    path: path.join(__dirname, 'public/'),
+    path: path.join(__dirname, 'dist/'),
   },
   entry: './src/index',
   plugins: [new RemoveEmptyScriptsPlugin()],
