@@ -1,11 +1,12 @@
 import path from 'path';
+import ansis from 'ansis';
 import { readDirRecursiveSync, readTextFileSync } from './file';
-import { compile } from './webpack';
-import { paths } from '../config';
+import { compile, watch } from './webpack';
+import { PATHS } from '../config';
 
 /**
  * This is the patch for some environments, like `jest`.
- * The `jest` hasn't in global scope the `btoa` function which used in `css-loader`.
+ * The `jest` hasn't in global scope the `btoa` function which is used in `css-loader`.
  */
 if (typeof global.btoa === 'undefined') {
   global.btoa = (input) => Buffer.from(input, 'latin1').toString('base64');
@@ -13,72 +14,143 @@ if (typeof global.btoa === 'undefined') {
 
 export const getCompareFileList = function (receivedPath, expectedPath) {
   return {
-    received: readDirRecursiveSync(receivedPath, false).sort(),
-    expected: readDirRecursiveSync(expectedPath, false).sort(),
+    received: readDirRecursiveSync(receivedPath)
+      .map((file) => path.relative(receivedPath, file))
+      .sort(),
+    expected: readDirRecursiveSync(expectedPath)
+      .map((file) => path.relative(expectedPath, file))
+      .sort(),
   };
 };
 
-export const getCompareFileContents = function (receivedFile, expectedFile, filter = /.(html|css|css.map|js|json|js.map)$/) {
-  return !filter || filter.test(receivedFile) && filter.test(expectedFile)
+export const getCompareFileContents = function (receivedFile, expectedFile, filter = /.(html|css|css.map|js|js.map)$/) {
+  return filter.test(receivedFile) && filter.test(expectedFile)
     ? { received: readTextFileSync(receivedFile), expected: readTextFileSync(expectedFile) }
     : { received: '', expected: '' };
 };
 
-export const compareFileListAndContent = (done, dirname, filter = /.(html|css|css.map|js|json|js.map)$/) => {
-  const absTestPath = path.join(paths.testSource, dirname),
-    webRootPath = path.join(absTestPath, paths.webRoot),
-    expectedPath = path.join(absTestPath, paths.expected);
+/**
+ * Compare the file list and content of files.
+ *
+ * @param {string} relTestCasePath The relative path to the test directory.
+ * @param {boolean} compareContent Whether the content of files should be compared too.
+ * @return {Promise<void>}
+ */
+export const compareFiles = (relTestCasePath, compareContent = true) => {
+  const absTestPath = path.join(PATHS.testSource, relTestCasePath),
+    webRootPath = path.join(absTestPath, PATHS.webRoot),
+    expectedPath = path.join(absTestPath, PATHS.expected);
 
-  compile(dirname, {}).then(() => {
-    const { received: receivedFiles, expected: expectedFiles } = getCompareFileList(webRootPath, expectedPath);
-    expect(receivedFiles).toEqual(expectedFiles);
+  return expect(
+    compile(PATHS, relTestCasePath, {})
+      .then(() => {
+        const { received: receivedFiles, expected: expectedFiles } = getCompareFileList(webRootPath, expectedPath);
+        expect(receivedFiles).toEqual(expectedFiles);
 
-    expectedFiles.forEach((file) => {
-      const { received, expected } = getCompareFileContents(
-        path.join(webRootPath, file),
-        path.join(expectedPath, file),
-        filter
-      );
-      expect(received).toEqual(expected);
-    });
-    done();
-  });
+        if (compareContent) {
+          expectedFiles.forEach((file) => {
+            const { received, expected } = getCompareFileContents(
+              path.join(webRootPath, file),
+              path.join(expectedPath, file)
+            );
+            expect(received).toEqual(expected);
+          });
+        }
+
+        return Promise.resolve(true);
+      })
+      .catch((error) => {
+        return Promise.reject(error);
+      })
+  ).resolves.toBe(true);
 };
 
-export const compareFileList = (done, dirname) => {
-  const absTestPath = path.join(paths.testSource, dirname),
-    webRootPath = path.join(absTestPath, paths.webRoot),
-    expectedPath = path.join(absTestPath, paths.expected);
+export const watchCompareFiles = (relTestCasePath) => {
+  const absTestPath = path.join(PATHS.testSource, relTestCasePath),
+    webRootPath = path.join(absTestPath, PATHS.webRoot),
+    expectedPath = path.join(absTestPath, PATHS.expected);
 
-  compile(dirname, {}).then(() => {
-    const { received: receivedFiles, expected: expectedFiles } = getCompareFileList(webRootPath, expectedPath);
-    expect(receivedFiles).toEqual(expectedFiles);
-    done();
-  });
+  return expect(
+    watch(PATHS, relTestCasePath, { devServer: { hot: true } })
+      .then(() => {
+        const { received: receivedFiles, expected: expectedFiles } = getCompareFileList(webRootPath, expectedPath);
+        expect(receivedFiles).toEqual(expectedFiles);
+
+        expectedFiles.forEach((file) => {
+          const { received, expected } = getCompareFileContents(
+            path.join(webRootPath, file),
+            path.join(expectedPath, file)
+          );
+          expect(received).toEqual(expected);
+        });
+        return Promise.resolve(true);
+      })
+      .catch((error) => {
+        return Promise.reject(error);
+      })
+  ).resolves.toBe(true);
 };
 
-export const exceptionContain = function (done, dirname, containString) {
-  compile(dirname, {})
-    .then(() => {
-      throw new Error('the test should throw an error');
+export const exceptionContain = (relTestCasePath, containString) => {
+  return expect(
+    compile(PATHS, relTestCasePath, {})
+      .then(() => {
+        return Promise.reject('the test should throw an error');
+      })
+      .catch((error) => {
+        const message = ansis.strip(error.toString());
+        return Promise.reject(message);
+      })
+  ).rejects.toContain(containString);
+};
+
+export const watchExceptionContain = function (relTestCasePath, containString) {
+  return expect(
+    watch(PATHS, relTestCasePath, {}, (watching) => {
+      watching.close();
     })
-    .catch((error) => {
-      expect(error.toString()).toContain(containString);
-      done();
-    });
+      .then(() => {
+        return Promise.reject('the test should throw an error');
+      })
+      .catch((error) => {
+        const message = ansis.strip(error.toString());
+        return Promise.reject(message);
+      })
+  ).rejects.toContain(containString);
 };
 
-export const stdoutContain = function (done, dirname, containString) {
-  const stdout = jest.spyOn(console._stdout, 'write').mockImplementation(() => {});
+export const stdoutSnapshot = function (relTestCasePath, done) {
+  const stdout = jest.spyOn(process.stdout, 'write').mockImplementation(() => {});
 
-  compile(dirname, {}).then(() => {
-    const { calls } = stdout.mock;
-    const output = calls.length > 0 ? calls[0][0] : '';
+  return expect(
+    compile(PATHS, relTestCasePath, {}).then(() => {
+      const { calls } = stdout.mock;
+      let output = calls.length > 0 ? calls[0][0] : '';
+      output = ansis.strip(output);
 
-    stdout.mockClear();
-    stdout.mockRestore();
+      stdout.mockClear();
+      stdout.mockRestore();
 
-    expect(output).toContain(containString);
-    done();
-  });
+      return Promise.resolve(output);
+    })
+  ).resolves.toMatchSnapshot();
+};
+
+export const watchStdoutSnapshot = function (relTestCasePath, done) {
+  const stdout = jest.spyOn(process.stdout, 'write').mockImplementation(() => {});
+
+  return expect(
+    watch(PATHS, relTestCasePath, {}, (watching) => {
+      watching.close();
+    }).then(() => {
+      const { calls } = stdout.mock;
+      let output = calls.length > 0 ? calls[0][0] : '';
+      output = ansis.strip(output);
+
+      stdout.mockClear();
+      stdout.mockRestore();
+
+      return Promise.resolve(output);
+    })
+  ).resolves.toMatchSnapshot();
 };
