@@ -1,6 +1,11 @@
 module.exports = {
+  devtool: false,
+  // avoid double error output in console
+  stats: 'errors-warnings',
+
   output: {
     filename: '[name].js',
+    // clean the output directory before emit
     clean: true,
   },
 
@@ -10,10 +15,10 @@ module.exports = {
     rules: [],
   },
 
-  optimization: {
-    removeEmptyChunks: true,
-    mergeDuplicateChunks: true,
-    usedExports: true,
-    concatenateModules: true,
-  },
+  // optimization: {
+  //   removeEmptyChunks: true,
+  //   mergeDuplicateChunks: true,
+  //   usedExports: true,
+  //   concatenateModules: true,
+  // },
 };
